pyproject.toml
[tool.poetry]
name = "parking-media-wall"
version = "0.1.0"
description = "FastAPI-based parking media wall with editable YAML config, media carousel, and upstream parking parser proxy"
authors = ["OpenAI <support@openai.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "^0.116.1"
uvicorn = { extras = ["standard"], version = "^0.35.0" }
httpx = "^0.28.1"
PyYAML = "^6.0.2"
jinja2 = "^3.1.6"
pydantic = "^2.11.7"
python-multipart = "^0.0.20"

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.9"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
